Extract management tests
test_api_utils contained both base tests cases and actual tests cases.
Extracted the real ones to a separate module. Those were Management
(health, metrics, versions) test cases.
